Improve shared lock implementation
During lock creation:
Shared locks now have a reservation directory named similar to the exclusive
directories, instead of just 'shared'.  This fixes a race condition where
a process giving up a lock could remove the 'shared' directory successfully.
Meanwhile another wants the shared lock, recreates 'shared', drops it's file
in there, sees the symlink already points to 'shared' and assummes it has
the lock.  Then the first process continues by removing the symlink.  Now
process 2 thinks it has the lock, but it really doesn't.

The fix is if there is any 'shared-*' directory we should use it, otherwise
make a new one.  Now, process 2 will see that the symlink exists, but does
not point to the reservation directory it created, and knows it does not have
the lock.

During unlock:
For exclusive locks, we do the safer method of removing the lock symlink
first, then remove the reservation directory.  That way if we crash before
removing the directory, the lock is still successfully freed.

Shared locks must use the unsafer method of attempting to remove the
reservation directory first, then the symlink of rmdir was successful.

Refuse to free the lock if the current PID does not match the PID used
during create